% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_gen.R
\name{accessory_gen}
\alias{accessory_gen}
\title{Extract the representative Accessory Gene sequences.}
\usage{
accessory_gen(inputFiles, organisms, outputFiles)
}
\arguments{
\item{inputFiles}{user defined name of previously clustered file with fasta format.}

\item{organisms}{A number.}

\item{outputFiles}{User defined name of the file with fasta extension}
}
\value{
Number of accessory gene sequences in outputFiles present in pan-genome consisting of \code{organisms} included in this pan-genome study.
}
\description{
Extract the representative Accessory Gene sequences.
}
\examples{
accessory_gen("clusters.fasta", 14, "access.fasta")
accessory_gen(inputFiles, 14, outputFile)
}
