% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessorygenes.R
\name{accessorygenes}
\alias{accessorygenes}
\title{Returns the Number of Accessory Gene sequences.}
\usage{
accessorygenes(file)
}
\arguments{
\item{file}{user defined name of previously binary clustered file with csv format.}
}
\value{
Number of accessory gene sequences in outputFiles present in pan-genome included in this pan-genome study.
}
\description{
Returns the Number of Accessory Gene sequences.
}
\examples{
accessorygenes("bin_matrix.csv")
}
