% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotat.R
\name{annotat}
\alias{annotat}
\title{Locally aligned the entered sequences with any given database.}
\usage{
annotat(inputFiles, database, evalue, id, outputFiles)
}
\arguments{
\item{inputFiles}{user defined name of sequences intended to be aligned with fasta format.}

\item{database}{user defined name of database with fasta format.}

\item{evalue}{A number e.g. 1E-6, 0.0 etc.}

\item{id}{A number ranging from 0.3 to 1.}

\item{outputFiles}{User defined name of the file with txt extension}
}
\value{
Locally aligned sequences with their annotations and other particulars in outputFiles.
}
\description{
Locally aligned the entered sequences with any given database.
}
\examples{
annotate("sequences.fasta", "keggdatabase.fasta", 1E-6, 0.4 ,"kegg.txt")
}
