% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_bin.R
\name{clust_bin}
\alias{clust_bin}
\title{Clusters all the genome sequences and returns binary pan-matrix.}
\usage{
clust_bin(inputFiles, id, outputFiles, y)
}
\arguments{
\item{inputFiles}{user defined name of previously clustered file with fasta format.}

\item{id}{A number ranging from 0.5 to 1.}

\item{outputFiles}{User defined name of the file with fasta extension}

\item{y}{total number of organisms included for pan-genomic study.}
}
\value{
Returns a binary pan-matrix file with csv extension e.g. bin_matrix.csv.
}
\description{
Clusters all the genome sequences and returns binary pan-matrix.
}
\examples{
clust_bin("sorted.fasta", 0.5, "clusters.fasta", 14)
}
