% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate.R
\name{concatenate}
\alias{concatenate}
\title{concatenate all the Gene/Genome sequences.}
\usage{
concatenate(V1, V2)
}
\arguments{
\item{V1}{user defined name of previously clustered file with fasta format.}

\item{V2}{user defined name of the output file.}
}
\value{
Concatenating all the gene/genome sequences in the user defined directory.
}
\description{
concatenate all the Gene/Genome sequences.
}
\examples{
for( i in 1:dim(genome.table)[1] ){
cat("Concatenating the sequences", genome.table$Files[i], "...\\n")
in.file <- file.path("../sequence", genome.table$Files[i])
out.file <- file.path("../concatenateseq", genome.table$Files[i])
Concatenate(in.file,out.file)}
}
