% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert all the fasta formatted Gene/Genome sequences to txt formats.}
\usage{
convert(old, new)
}
\arguments{
\item{old}{all the files in the directory with fasta/txt format.}

\item{new}{all the files in the directory with fasta/txt format.}
}
\value{
convert all the gene/genome sequences in the user defined directory to txt/fasta.
Need to set the working directory everytime.
}
\description{
Convert all the fasta formatted Gene/Genome sequences to txt formats.
}
\examples{
convert("txt", "fasta")
convert("fasta", "txt")
}
