% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coregenes.R
\name{coregenes}
\alias{coregenes}
\title{Returns the Number of Core Gene sequences.}
\usage{
coregenes(file)
}
\arguments{
\item{file}{user defined name of previously binary clustered file with csv format.}
}
\value{
Number of core gene sequences in outputFiles present in pan-genome included in this pan-genome study.
}
\description{
Returns the Number of Core Gene sequences.
}
\examples{
coregenes("bin_matrix.csv")
}
