% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distGen.R
\name{distGen}
\alias{distGen}
\title{Generate a distance matrix table based on the binary pan-matrix.}
\usage{
distGen(file)
}
\arguments{
\item{file}{user defined name of previously clustered file with csv format e.g. bin_matrix.csv.}
}
\value{
Returns a binary pan-matrix file with csv extension e.g. ID_matrix.csv.
}
\description{
Generate a distance matrix table based on the binary pan-matrix.
}
\examples{
distGen("bin_matrix.csv")
}
