% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatGen1.R
\name{heatgen1}
\alias{heatgen1}
\title{Generate a Denrogram and heatmap based on the binary pan-matrix.}
\usage{
heatgen1(file)
}
\arguments{
\item{file}{user defined name of previously clustered file with csv format e.g. bin_matrix.csv.}
}
\value{
Returns a dendrogram and heatmap.
}
\description{
Generate a Denrogram and heatmap based on the binary pan-matrix.
}
\examples{
heatgen1("bin_matrix.csv")
}
