% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatGen2.R
\name{heatgen2}
\alias{heatgen2}
\title{Generate heatmap based on presence/absence of gene.}
\usage{
heatgen2(file)
}
\arguments{
\item{file}{user defined name of previously clustered file with csv format e.g. bin_matrix.csv.}
}
\value{
Returns a heatmap based on presence/absence of gene.
}
\description{
Generate heatmap based on presence/absence of gene.
}
\examples{
heatgen2("bin_matrix.csv")
}
