% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinfasta.R
\name{joinfasta}
\alias{joinfasta}
\title{Join all the fasta formatted Gene/Genome sequences in the working directory.}
\usage{
joinfasta(allfasta)
}
\arguments{
\item{allfasta}{write all fasta to join the all fasta sequences present in the directory.}
}
\value{
join all the gene/genome sequences in the user defined directory and returns combined.fasta.
Need to set the working directory everytime.
}
\description{
Join all the fasta formatted Gene/Genome sequences in the working directory.
}
\examples{
convert("txt", "fasta")
convert("fasta", "txt")
}
