% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangen.R
\name{panGen}
\alias{panGen}
\title{Enumerates the Gene sequences present in all the genomes.}
\usage{
panGen(file)
}
\arguments{
\item{file}{user defined name of previously clustered file with fasta format.}
}
\value{
enumerates all the gene sequences in outputFiles present in pan-genome.
}
\description{
Enumerates the Gene sequences present in all the genomes.
}
\examples{
panGen("bin_matrix.csv")
}
