% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predORFaa.R
\name{predORFaa}
\alias{predORFaa}
\title{Predicts and writes the amino acid sequences of ORF in Genome sequences.}
\usage{
predORFaa(inputFiles, outputFiles)
}
\arguments{
\item{inputFiles}{user defined name of previously clustered file with fasta format.}

\item{outputFiles}{User defined name of the file with fasta extension}
}
\value{
Predicts the Amino acid sequences in outputFiles present in pan-genome.
}
\description{
Predicts and writes the amino acid sequences of ORF in Genome sequences.
}
\examples{
predORFaa("anyseq.fasta", "AA.fasta")
}
