% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predORFnuc.R
\name{predORFnuc}
\alias{predORFnuc}
\title{Predicts and writes the nucleotide sequences of ORF in Genome sequences.}
\usage{
predORFnuc(inputFiles, outputFiles)
}
\arguments{
\item{inputFiles}{user defined name of previously clustered file with fasta format.}

\item{outputFiles}{User defined name of the file with fasta extension}
}
\value{
Predicts the nucleotide ORF sequences in outputFiles.
}
\description{
Predicts and writes the nucleotide sequences of ORF in Genome sequences.
}
\examples{
predORFnuc("anyseq.fasta", "NUC.fasta")
}
