% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{relabel}
\alias{relabel}
\title{Relabel the Gene/Genome sequences.}
\usage{
relabel(inputFiles, label, outputFiles)
}
\arguments{
\item{inputFiles}{user defined name of previously clustered file with fasta format.}

\item{label}{A new number with organims id and gene e.g. org1|gene.}

\item{outputFiles}{User defined name of the file with fasta extension}
}
\value{
Relabels gene/genome sequences in outputFiles.
}
\description{
Relabel the Gene/Genome sequences.
}
\examples{
relabel("organism.fasta", organism|geneno, "relabel.fasta")
for( i in 1:dim(genome.table)[1] ){
cat("Relabelling the fasta sequences", genome.table$Files[i], "...\\n")
in.file <- file.path("../concatenate", genome.table$Files[i])
label <- file.path(genome.table$Strains[i])
out.file <- file.path("../relabel", genome.table$Files[i])
relabel(in.file,label,out.file)}
}
