% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortbylength.R
\name{sortbylength}
\alias{sortbylength}
\title{Sort the Gene/Genome sequences by length (A necessary step before clustering).}
\usage{
sortbylength(inputFiles, outputFiles)
}
\arguments{
\item{inputFiles}{user defined name of previously clustered file with fasta format.}

\item{outputFiles}{User defined name of the file with fasta extension}
}
\value{
Sorts the gene/genome sequences by length in outputFiles.
}
\description{
Sort the Gene/Genome sequences by length (A necessary step before clustering).
}
\examples{
Sortbylength("combined.fasta", "sorted.fasta")
}
