% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniq_gen.R
\name{uniq_gen}
\alias{uniq_gen}
\title{Extract the representative Unique Gene sequences.}
\usage{
uniq_gen(inputFiles, uniq, outputFiles)
}
\arguments{
\item{inputFiles}{user defined name of previously clustered file with fasta format.}

\item{uniq}{Least number of organisms e.g. 1.}

\item{outputFiles}{User defined name of the file with fasta extension}
}
\value{
Number of unique gene sequences in outputFiles present in pan-genome consisting of \code{organisms} included in this pan-genome study.
}
\description{
Extract the representative Unique Gene sequences.
}
\examples{
uniq_gen("clusters.fasta", 1, "uniq.fasta")
uniq_gen(inputFiles, 1, outputFile)
}
