% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uniqgenes.R
\name{Uniqgenes}
\alias{Uniqgenes}
\title{Returns the Number of Unique Gene sequences.}
\usage{
Uniqgenes(file)
}
\arguments{
\item{file}{user defined name of previously binary clustered file with csv format.}
}
\value{
Number of accessory gene sequences in outputFiles present in pan-genome included in this pan-genome study.
}
\description{
Returns the Number of Unique Gene sequences.
}
\examples{
Uniqgenes(file)
}
